pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "tegrastats-api"
version = "1.0.0"
description = "HTTP API and WebSocket server for NVIDIA Jetson tegrastats monitoring"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Tegrastats API Team", email = "contact@example.com"}
]
maintainers = [
    {name = "Tegrastats API Team", email = "contact@example.com"}
]
keywords = ["jetson", "tegrastats", "monitoring", "api", "websocket", "esp32"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Hardware",
    "Topic :: System :: Monitoring",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
]
requires-python = ">=3.8"
dependencies = [
    "Flask>=2.3.0",
    "Flask-SocketIO>=5.3.0",
    "Flask-CORS>=4.0.0",
    "python-socketio>=5.9.0",
    "eventlet>=0.33.0",
    "psutil>=5.9.0",
    "requests>=2.31.0",
    "websocket-client>=1.6.0",
    "click>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
test = [
    "pytest>=7.0.0",
    "requests>=2.31.0",
    "websocket-client>=1.6.0",
]

[project.urls]
Homepage = "https://github.com/your-org/tegrastats-api"
Documentation = "https://github.com/your-org/tegrastats-api/blob/main/README.md"
Repository = "https://github.com/your-org/tegrastats-api.git"
"Bug Tracker" = "https://github.com/your-org/tegrastats-api/issues"

[project.scripts]
tegrastats-api = "tegrastats_api.cli:main"
tegrastats-server = "tegrastats_api.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
tegrastats_api = ["templates/*", "static/*", "config/*.py"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true